Rename the viaduct Android package to `viaduct`.
Before it was named `httpconfig` for historical reasons, this one seems
more natural.
