+++ components/viaduct/android/build.gradle
@@ -2,7 +2,7 @@ apply from: "$appServicesRootDir/build-scripts/component-common.gradle"
 apply from: "$appServicesRootDir/publish.gradle"
 
 android {
-    namespace 'org.mozilla.appservices.httpconfig'
+    namespace 'org.mozilla.appservices.viaduct'
 }
 
 dependencies {
