+++ components/viaduct/android/src/main/java/mozilla/appservices/viaduct/HttpConfig.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.appservices.httpconfig
+package mozilla.appservices.viaduct
 
 import mozilla.appservices.viaduct.initBackend
 import mozilla.components.concept.fetch.Client
